Better list-games

+++ libs/bson-models/src/publisher.ts
@@ -2,7 +2,7 @@ import type { PostgresRecordModel } from 'from-schema';
 
 export const publisher = {
 	properties: {
-		id: { type: 'bigint', primaryKey: true },
+		id: { type: 'integer', primaryKey: true },
 		homepage: { type: 'text', minLength: 1, maxLength: 50 },
 		name: { type: 'text', minLength: 1, maxLength: 50 },
 		created: { type: 'timestamptz', default: { sql: 'CURRENT_TIMESTAMP' } },
